added "non-null ref"
"non-null ref" (`NonNullRef`) is a Giter that yields a sequence of references to elements (originally pointers)
It effectively represents the following:
```cpp
auto q = s
    | Where([](const int* p) { return p != nullptr; })
    | Select([](int* p) -> int& { return *p; });
```

`Select` has also been updated to allow the long-hand form shown above.
This required that the return type for `op*` be explicit (so as to preserve reference types).

There appears to be a bug in optimized builds :(
Something like... I'm using references to temp objects that are going out of scope?